ingest/requirements.txt
aiohttp
pyyaml
duckdb
pandas
